Decompile dtTick

+++ src/include/game/data/data_02da90.h
@@ -32,7 +32,7 @@ extern u32 misc_bios[];
 extern u32 var80088964;
 extern u32 place_bios[];
 extern u8 var80088ad8;
-extern u32 var80088adc;
+extern u8 var80088adc;
 extern u32 var80088ae0;
 extern u32 var80088aec;
 extern u32 g_CiDeviceStageFlags[10];
